scripts/requirements.txt
matplotlib
pandas 
tqdm
tensorflow_addons
tensorflow_datasets
imageio
seaborn
jinja2
scipy
scikit-learn
opencv-python
openpyxl
umap-learn